# implementations/java/src/main/java/com/eyex/api/model/SimulateBatchResponse.java
package com.eyex.api.model;

import java.util.List;

public record SimulateBatchResponse(
        String type,
        double severity,
        String model,
        List<SimulatedColor> results
) {}